Api 34330 split controllers (#15787)
* Split controllers

* Simplify route config a little

* Another small cleanup

* Remove extra route

* Split spec also

* Move format_org

---------
